% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defIgRanges.R
\name{defIgRanges}
\alias{defIgRanges}
\title{Define intergenic genomic regions}
\usage{
defIgRanges(fgtf, genome = NULL, fchromsize = NULL, radius = 10000,
  feat = "exon")
}
\arguments{
\item{fgtf}{Name of the GTF file to define genic regions. Required to have
gene_id in the attribute column (column 9).}

\item{genome}{Version of the genome. Currently supported ones are:
\itemize{
    \item hg19
    \item hg38
    \item mm9
    \item mm10
}
All the above genomes have sizes for major chromosomes, i.e.
non-random, non-alt, or mitochondria chromosomes. For hg19
and hg38, it contains chr1-22, X, and Y. For mm9 and mm10,
it contains chr1-19, X, and Y.
Default: NULL}

\item{fchromsize}{Name of a file defining chromosome sizes as a supplement
to unsupported genomes. Often, it can be downloaded from
UCSC, e.g. for hg19, http://hgdownload.cse.ucsc.edu/goldenpath/hg19/database/chromInfo.txt.gz
Required to have at least two tab-delimited columns
without any header:
\enumerate{
   \item chromosome name, e.g. chr1
   \item chromosome length, e.g. 249250621
}
Both uncompressed and gzipped files are supported.
Default: NULL}

\item{radius}{Region length (bp) of gene's upstream and downstream to be
excluded from intergenic region.
Default: 10,000}

\item{feat}{feature in the GTF file to-be-used for defining genic region.
Default: exon}
}
\value{
a GRanges object of intergenic regions
}
\description{
Define intergenic genomic regions
}
