% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMdl.R
\name{evalMdl}
\alias{evalMdl}
\title{Evaluate transcript model}
\usage{
evalMdl(mdlexondt, tgtexondt)
}
\arguments{
\item{mdlexondt}{a data.table of model exons with columns: chrom, start,
end, strand, and transcript ID.}

\item{tgtexondt}{a data.table of target exons with columns: chrom, start,
end, strand, and transcript ID.}
}
\value{
a data table of precision, recall, number of true positive,
        false negative, false positive for all three evaluated features
}
\description{
Evaluate trascript model's precision and recall on exon nucleotide,
splice junctions, and transcript's splice pattern
}
