% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMdl.R
\name{evalMdlExonNuc}
\alias{evalMdlExonNuc}
\title{Evaluate transcript model's exon nucleotide}
\usage{
evalMdlExonNuc(mdldt, tgtdt)
}
\value{
a data.table of precision, recall, number of true positive,
}
\description{
Evaluate transcript model's exon nucleotide
}
