% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMdl.R
\name{evalMdlJnc}
\alias{evalMdlJnc}
\title{Evaluate transcript model's splice junction}
\usage{
evalMdlJnc(tgtjnc_in_mdldt, mdljnc_in_tgtdt, mdl_ol_tgtdt, tgtjncdt)
}
\value{
a data table of precision, recall, number of true positive,
        false negative, false positive at each junction and all junctions in
        a transcript
}
\description{
Evaluate transcript model's splice junction
}
