% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalModel.R
\name{evalModel}
\alias{evalModel}
\title{Evaluate transcript model}
\usage{
evalModel(mdltr, tgttr)
}
\arguments{
\item{mdltr}{a Transcript object for models}

\item{tgttr}{a Transcript object for targets}
}
\value{
a data table of precision, recall, number of true positive,
        false negative, false positive for all three evaluated features
}
\description{
Evaluate trascript model's precision and recall on exon nucleotide,
splice junctions, and transcript's splice pattern
}
