% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{filterBam4IG}
\alias{filterBam4IG}
\title{filter bam files for a given intergenic regions}
\usage{
filterBam4IG(fbams, iggrs, nthr, tmpdir)
}
\value{
a data.table contains input and filtered bam files by chromosome and
        strand of a given intergenic genomic ranges
}
\description{
filter bam files for a given intergenic regions
}
