% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalModel.R
\name{findModelJncInTarget}
\alias{findModelJncInTarget}
\title{find if model's splice junction exists in a target transcript}
\usage{
findModelJncInTarget(mdljncdt, tgtjncdt)
}
\description{
find if model's splice junction exists in a target transcript
}
