% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcript.R
\name{getTrFromExon}
\alias{getTrFromExon}
\title{get transcript genomic ranges from exons}
\usage{
getTrFromExon(exondt, id_col_name = "trid")
}
\arguments{
\item{exondt}{a data table of exon with columns: chrom, start, end, strand,
and transcript ID}

\item{id_col_name}{the column name of transcript ID, default is 'trid'}
}
\value{
a data table of transcript genomic ranges with columns: chrom, start,
        end, strand, number of exons, and transcript ID
}
\description{
get transcript genomic ranges from exons
}
