% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcript.R
\name{getTrJncFromExon}
\alias{getTrJncFromExon}
\title{obtain junction info from exon}
\usage{
getTrJncFromExon(exondt)
}
\arguments{
\item{exondt}{a data.table of exons with columns: chrom, start, end, strand,
and transcript ID}
}
\value{
a data table of junctions with columns: chrom, start, end, strand,
        transcript ID, number of junctions, and junction's index in
        transcript
}
\description{
obtain junction info from exon
}
