% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTF.R
\docType{methods}
\name{initFromDataTable,GTF,data.table,vector-method}
\alias{initFromDataTable,GTF,data.table,vector-method}
\title{construct a GTF object from a data.table object}
\usage{
\S4method{initFromDataTable}{GTF,data.table,vector}(obj, dt, infokeys, ...)
}
\arguments{
\item{obj}{a GTF object to be initialized}

\item{dt}{a data.table object to define genomic ranges}

\item{infokeys}{a vector of characters to define to-be-extracted entries
in GTF file's column 9}
}
\description{
construct a GTF object from a data.table object
}
