% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTF.R
\docType{methods}
\name{initFromGRanges,GRanges-method}
\alias{initFromGRanges,GRanges-method}
\title{construct a GTF object from a GenomicRanges object}
\usage{
\S4method{initFromGRanges}{GRanges}(grs)
}
\description{
construct a GTF object from a GenomicRanges object
}
