% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTF.R
\docType{methods}
\name{initFromGRanges,GTF,GRanges-method}
\alias{initFromGRanges,GTF,GRanges-method}
\title{construct a GTF object from a GenomicRanges object}
\usage{
\S4method{initFromGRanges}{GTF,GRanges}(obj, grs)
}
\arguments{
\item{obj}{a GTF object to be initialized}

\item{grs}{a GenomicRanges object to define ranges in GTF file}
}
\description{
construct a GTF object from a GenomicRanges object
}
