% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTF.R
\docType{methods}
\name{initFromGTFFile,GTF,character,vector-method}
\alias{initFromGTFFile,GTF,character,vector-method}
\title{initialize a GTF object from a GTF file with a vector of info keys to parse}
\usage{
\S4method{initFromGTFFile}{GTF,character,vector}(obj, fgtf, infokeys, ...)
}
\arguments{
\item{obj}{a GTF object to be initialized}

\item{fgtf}{file name with full path to a GTF file}

\item{infokeys}{a vector of characters to define to-be-extracted entries in GTF file's column 9}
}
\description{
initialize a GTF object from a GTF file with a vector of info keys to parse
}
