% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcript.R
\docType{methods}
\name{initialize,Transcript-method}
\alias{initialize,Transcript-method}
\title{Construct a Transcript Object}
\usage{
\S4method{initialize}{Transcript}(.Object, exon, ...)
}
\arguments{
\item{exon}{a data.table object of exon}
}
\value{
a Transcript object
}
\description{
Construct a transcript object from an exon data.table
}
