% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepIgBam.R
\name{prepIgBam}
\alias{prepIgBam}
\title{Prepare BAM files for intergenic regions}
\usage{
prepIgBam(fbams, iggrs, outdir, nthreads = 1)
}
\arguments{
\item{fbams}{A vector of characters for a list of input RNA-seq BAM files.
Currently, PRAM only supports strand-specific paired-end
data with the first mate on the right-most of transcript
coordinate, i.e., 'fr-firststrand' by Cufflinks's
definition}

\item{iggrs}{A GenomicRanges object defining intergenic regions}

\item{outdir}{A string defining the full name of an directory to
save all the output files}

\item{nthreads}{An integer defining the number of threads. Default: 1}
}
\description{
Prepare BAM files for intergenic regions
}
