% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPRAM.R
\name{runPRAM}
\alias{runPRAM}
\title{Predict intergenic transcript models from RNA-seq and screen them by ChIP-seq}
\usage{
runPRAM(in_gtf, in_bamv, out_gtf, cufflinks, genome, in_bedv, training_gtf,
  training_tpms)
}
\arguments{
\item{in_gtf}{An input GTF file for defining genomic coordinates of
existing
genes.  Required to have `gene_id` in the attribute column
(column 9)}

\item{in_bamv}{A character vector of input BAM file(s). If mode 'cf'
or 'st' is used, only one input RNA-seq BAM file is allowed.
Currently, PRAM only supports strand-specific paired-end data
with the first mate on the right-most of transcript
coordinate, i.e., 'fr-firststrand' by Cufflinks's definition.}

\item{out_gtf}{An output GTF file of predicted transcript models}

\item{cufflinks}{Cufflinks executable.  Required by mode 'plcf',
'cfmg', and 'cf'.  For mode 'cfmg', executable files of
Cuffmerge, Cuffcompare, and gtf_to_sam from the Cufflinks
suite are assumed to be under the same folder as
Cufflinks.
All the executables are available to download for
Linux \url{http://cole-trapnell-lab.github.io/cufflinks/assets/downloads/cufflinks-2.2.1.Linux_x86_64.tar.gz}
and MacOS \url{http://cole-trapnell-lab.github.io/cufflinks/assets/downloads/cufflinks-2.2.1.OSX_x86_64.tar.gz}
Default: ''}

\item{genome}{Version of the genome. Currently supported ones are:
\itemize{
    \item hg19
    \item hg38
    \item mm9
    \item mm10
}
All the above genomes have sizes for all chromosomes
including random and alt ones.
Default: NULL}

\item{in_bedv}{A vector of BED file(s) for histone marks and/or RNA Pol II
ChIP-seq alignments.  Each BED file is required to have at least
three columns describing chromosome names, starting and
ending positions. Do not include header in BED files.}

\item{training_gtf}{A GTF file that defines the genomic coordinates and features
for all training transcripts.  This file is required to have 'exon'
in the feature column (the 3rd column) and 'transcript_id' entry in
the attribute column (the 9th column).}

\item{training_tpms}{A vector of RSEM quantification result file(s) at the
isoform level.  It is required to have two columns named
'transcript_id' and 'TPM' to define the abundances of
training transcripts.  This file is the one with a suffix
'.isoforms.results' generated by RSEM. Header is required.}
}
\description{
Predict intergenic transcript models from RNA-seq and screen them by ChIP-seq
}
