% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bam.R
\name{selAlnByMateMaxNDup}
\alias{selAlnByMateMaxNDup}
\title{filter bam files for a given intergenic regions}
\usage{
selAlnByMateMaxNDup(alns, max_uni_ndup, max_mul_ndup)
}
\value{
a data.table contains input and filtered bam files by chromosome and
        strand of a given intergenic genomic ranges
}
\description{
filter bam files for a given intergenic regions
}
