% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenModel.R
\name{trainModelClassifier}
\alias{trainModelClassifier}
\title{Train a random forest classifier to filter transcript models}
\usage{
trainModelClassifier(fbeds, ftpms, fgtf, tmpdir = NULL, expr_min_tpm = 1,
  cv_n_folds = 10, nthreads = 1)
}
\arguments{
\item{fgtf}{A string of a GTF file defining the genomic coordinates for
all training transcripts.  The file is required for have 'exon'
in the feature column, where transcript coordinates will be
derived.}

\item{tmpdir}{Name of a temporary directory to save four files:
\itemize{
    \item 'rf.rda' for an R random forest classifier
    \item 'roc_pr.pdf' plots ROC and precision-recall curves
          from cross validation
    \item 'cv.tsv' saved the prediction results on training
          transcript from cross validation.
    \item 'pred.tsv' saved the prediction results testing transcripts
}
File with the same name in this folder will be overwritten.
Default: NULL}

\item{expr_min_tpm}{A numeric value of minimium TPM for a transcript to be defined as
expressed.  An 'expressed' or 'unexpressed'
transcript must have TPM >= or < this cutoff in all
supplied RSEM quantification files (ftpms),
respectively.  Otherwise, a transcript's expression
status will be defined as 'indeterminate'.  Transcripts with
'indeterminate' expression status will be omitted from the
training and output files.
Default: 1}

\item{cv_n_folds}{An integer of folds for cross validation. 'Expressed' and
'unexpressed' training transcripts will be stratified
for calculating AUC.
Default: 10}

\item{nthreads}{An integer defining the number of threads to-be-used.
An ideal number would be equal to the maximum one of
cv_n_folds and number of input BED files. Default: 1}
}
\value{
A list contains:
         \itemize{
             \item 'rf': a trained random forest object
             \item 'beddt': a data.table of processed ChIP-seq data
         }
}
\description{
Train a random forest classifier to filter transcript models
}
