% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTF.R
\docType{methods}
\name{writeGTF,GTF,character,logical-method}
\alias{writeGTF,GTF,character,logical-method}
\title{write a GTF object to a GTF file}
\usage{
\S4method{writeGTF}{GTF,character,logical}(x, fout, append)
}
\arguments{
\item{x}{a GTF object}

\item{fout}{a character object of GTF file name}

\item{append}{a boolean to inidicate if to append to a GTF file or not}
}
\description{
write a GTF object to a GTF file
}
