% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build transcript models from aligned RNA-seq data}
\usage{
buildModel(fbams, outdir, mode = "plcf", nthreads = 1, cufflinks = "",
  stringtie = "", taco = "")
}
\arguments{
\item{fbams}{A character vector of input RNA-seq BAM file(s).  Currently,
PRAM only supports strand-specific paired-end data with the
first mate on the right-most of transcript coordinate, i.e.,
'fr-firststrand' by Cufflinks's definition}

\item{outdir}{A character string defining the full name of a directory for
saving output files. PRAM will a folder named
'pram_tmp_$mode/' under this directory to save temporary
files.}

\item{mode}{A character string defining PRAM's model building mode.
Current available modes are:
\itemize{
    \item plcf: pooling   + cufflinks
    \item plst: pooling   + stringtie
    \item cfmg: cufflinks + cuffmerge
    \item stmg: stringtie + merging
    \item cftc: cufflinks + taco
    \item cf:   cufflinks
    \item st:   stringtie
}
Default: 'plcf'}

\item{nthreads}{An integer defining the number of threads to-be-used.
Default: 1}

\item{cufflinks}{Cufflinks executable file.  Required by mode 'plcf',
'cfmg', and 'cf'.  For mode 'cfmg', executable files of
Cuffmerge, Cuffcompare, and gtf_to_sam from the Cufflinks
suite are assumed to be under the same folder as
Cufflinks.  Default: ''}

\item{stringtie}{StringTie executable file.  Required by mode
'plst', 'stmg', and 'st'. Default: ''}

\item{taco}{TACO executable file. Required by mode 'cftc'.
Default: ''}
}
\description{
Build transcript models from aligned RNA-seq data
}
