% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build transcript models from aligned RNA-seq data}
\usage{
buildModel(fbams, outdir, mode = "pooling+cufflinks", nthreads = 1,
  cufflinks = "", stringtie = "", taco = "")
}
\arguments{
\item{fbams}{a character vector of input RNA-seq BAM files.  Currently,
PRAM only supports strand-specific paired-end data with the
first mate on the right-most of transcript coordinate, i.e.,
'fr-firststrand' by Cufflinks's definition}

\item{outdir}{a character string defining the full name of a directory for
saving output files. PRAM will a folder named 'pram_tmp/'}

\item{mode}{a character string defining PRAM's model building mode.
Must be one of 'pooling+cufflinks', 'pooling+stringtie',
'cufflinks+cuffmerge', 'stringtie+merging', or
'cufflinks+taco'.  Default: 'pooling+cufflinks'}

\item{cufflinks}{Cufflinks executable file. Required by mode
'pooling+cufflinks' and 'cufflinks+cuffmerge'.
For mode 'cufflinks+cuffmerge',  executable files of
Cuffmerge, Cuffcompare, and gtf_to_sam from the Cufflinks
suite will be assumed in the same folder as Cufflinks.
Default: ''}

\item{stringtie}{StringTie executable file. Required by mode
'pooling+stringtie' and 'stringtie+merging'. Default: ''}

\item{taco}{TACO executable file. Required by mode
'cufflinks+taco'. Default: ''}

\item{nthr}{an integer defining the number of threads to-be-used.
Default: 1}
}
\description{
Build transcript models from aligned RNA-seq data
}
