% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalmdl.R
\name{evalmdl}
\alias{evalmdl}
\title{Evaluate model's accuracy}
\usage{
evalmdl(fmdlgtf, ftgtgtf)
}
\arguments{
\item{fmdlgtf}{full name of model's GTF}

\item{ftgtgtf}{full name of true transcript's GTF}
}
\description{
Evaluate model's accuracy
}
