% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getUniChromOriDt}
\alias{getUniChromOriDt}
\title{Get a data.table of unique combinations of chrom and ori from input GRanges}
\usage{
getUniChromOriDt(grs)
}
\description{
Get a data.table of unique combinations of chrom and ori from input GRanges
}
