% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepIgBam.R
\name{prepIgBam}
\alias{prepIgBam}
\title{Prepare BAM files for intergenic regions}
\usage{
prepIgBam(fbams, iggrs, outdir, nthreads = 1, max_uni_n_dup_aln = 10,
  max_mul_n_dup_aln = 10)
}
\arguments{
\item{fbams}{A vector of characters for a list of input RNA-seq BAM files.
Currently, PRAM only supports strand-specific paired-end
data with the first mate on the right-most of transcript
coordinate, i.e., 'fr-firststrand' by Cufflinks's
definition}

\item{iggrs}{A GenomicRanges object defining intergenic regions}

\item{outdir}{A string defining the full name of an directory to
save all the output files}

\item{nthreads}{An integer defining the number of threads. Input BAM files
will be distributed by this number for processing.
Default: 1}

\item{max_uni_n_dup_aln}{Maximum number of uniquely mapped fragments
to report per each alignment.
Default: 10}

\item{max_mul_n_dup_aln}{Maximum number of multi-mapping fragments
to report per each alignment.
Default: 10}
}
\description{
Prepare BAM files for intergenic regions
}
