% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTF.R
\name{readGTF}
\alias{readGTF}
\title{parse a GTF file}
\usage{
readGTF(fgtf, info_keys)
}
\arguments{
\item{fgtf}{input GTF file}

\item{info_keys}{a vector of characters for attribute to be extracted
from GTF file's 9th column}
}
\value{
a data.table object
}
\description{
parse a GTF file
}
