% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPRAM.R
\name{runPRAM}
\alias{runPRAM}
\title{Predict intergenic transcript models from RNA-seq and screen them by ChIP-seq}
\usage{
runPRAM(in_gtf, in_bamv, out_gtf, in_bedv, training_tpms, training_gtf)
}
\arguments{
\item{in_gtf}{An input GTF file for defining genomic coordinates of
existing
genes.  Required to have `gene_id` in the attribute column
(column 9)}

\item{in_bamv}{A character vector of input BAM file(s). If mode 'cf'
or 'st' is used, only one input RNA-seq BAM file is allowed.
Currently, PRAM only supports strand-specific paired-end data
with the first mate on the right-most of transcript
coordinate, i.e., 'fr-firststrand' by Cufflinks's definition.}

\item{out_gtf}{An output GTF file of predicted transcript models}

\item{in_bedv}{A vector of BED file(s) for histone marks and/or RNA Pol II
ChIP-seq alignments.  Each BED file is required to have at least
three columns describing chromosome names, starting and
ending positions. Do not include header in BED files.}

\item{training_tpms}{A vector of RSEM quantification result file(s) at the
isoform level.  It is required to have two columns named
'transcript_id' and 'TPM' to define the abundances of
training transcripts.  This file is the one with a suffix
'.isoforms.results' generated by RSEM. Header is required.}

\item{training_gtf}{A GTF file that defines the genomic coordinates and features
for all training transcripts.  This file is required to have 'exon'
in the feature column (the 3rd column) and 'transcript_id' entry in
the attribute column (the 9th column).}
}
\examples{

in_gtf = system.file('extdata/demo/in.gtf', package='pram')

in_bamv = c( system.file('extdata/demo/SZP.bam', package='pram'),
             system.file('extdata/demo/TLC.bam', package='pram') )

pred_out_gtf = tempfile(fileext='.gtf')

runPRAM(in_gtf, in_bamv, pred_out_gtf)


in_bedv = c( system.file('extdata/demo/H3K79me2.bed.gz', package='pram'),
             system.file('extdata/demo/POLR2.bed.gz',    package='pram') )

training_tpms = c( system.file('extdata/demo/AED1.isoforms.results',
                               package='pram'),
                   system.file('extdata/demo/AED2.isoforms.results',
                               package='pram') )

training_gtf = system.file('extdata/demo/training.gtf', package='pram')

screen_out_gtf = tempfile(fileext='.gtf')

runPRAM(in_gtf, in_bamv, screen_out_gtf, in_bedv, training_tpms,
        training_gtf)


}
