% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPRAM.R
\name{runPRAM}
\alias{runPRAM}
\title{Predict intergenic transcript models from RNA-seq}
\usage{
runPRAM(in_gtf, in_bamv, out_gtf)
}
\arguments{
\item{in_gtf}{An input GTF file for defining genomic coordinates of
existing
genes.  Required to have `gene_id` in the attribute column
(column 9)}

\item{in_bamv}{A character vector of input BAM file(s). If mode 'cf'
or 'st' is used, only one input RNA-seq BAM file is allowed.
Currently, PRAM only supports strand-specific paired-end data
with the first mate on the right-most of transcript
coordinate, i.e., 'fr-firststrand' by Cufflinks's definition.}

\item{out_gtf}{An output GTF file of predicted transcript models}
}
\value{
None
}
\description{
Predict intergenic transcript models from RNA-seq
}
\examples{

in_gtf = system.file('extdata/demo/in.gtf', package='pram')

in_bamv = c(system.file('extdata/demo/SZP.bam', package='pram'),
            system.file('extdata/demo/TLC.bam', package='pram') )

pred_out_gtf = tempfile(fileext='.gtf')

## if you are on Linux or MacOS, run the following:
# runPRAM(in_gtf, in_bamv, pred_out_gtf)


}
