% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenModel.R
\name{screenModel}
\alias{screenModel}
\title{Screen transcript models by a random forest classifier trained on
        known transcripts}
\usage{
screenModel(in_bedv, training_tpms, training_gtf, testing_gtf, out_gtf,
  tmpdir = NULL, expr_min_tpm = 1, cv_n_folds = 10, nthreads = 1)
}
\arguments{
\item{in_bedv}{A vector of BED file(s) for histone marks and/or RNA Pol II
ChIP-seq alignments.  Each BED file is required to have at least
three columns describing chromosome names, starting and
ending positions. Do not include header in BED files.}

\item{training_tpms}{A vector of RSEM quantification result file(s) at the
isoform level.  It is required to have two columns named
'transcript_id' and 'TPM' to define the abundances of
training transcripts.  This file is the one with a suffix
'.isoforms.results' generated by RSEM. Header is required.}

\item{training_gtf}{A GTF file that defines the genomic coordinates and features
for all training transcripts.  This file is required to have 'exon'
in the feature column (the 3rd column) and 'transcript_id' entry in
the attribute column (the 9th column).}

\item{testing_gtf}{Name of a GTF file defining the genomic coordiantes and features
for transcript models to be screened.  GTF files generated by
PRAM's buildModel() function can be used here directly.}

\item{out_gtf}{Name of a GTF file to save transcripts selected by the classifier}

\item{tmpdir}{Name of a temporary directory to save four files:
\itemize{
    \item 'rf.rda' for an R random forest classifier
    \item 'roc_pr.pdf' plots ROC and precision-recall curves
          from cross-validation
    \item 'cv.tsv' saved the prediction results on training
          transcript from cross-validation.
    \item 'pred.tsv' saved the prediction results testing transcripts
}
File with the same name in this folder will be overwritten.
Default: NULL}

\item{expr_min_tpm}{A numeric value of minimium TPM for a transcript to be defined as
expressed.  An 'expressed' or 'unexpressed'
transcript must have TPM >= or < this cutoff in all
supplied RSEM quantification files (ftpms),
respectively.  Otherwise, a transcript's expression
status will be defined as 'indeterminate'.  Transcripts with
'indeterminate' expression status will be omitted from the
training and output files.
Default: 1}

\item{cv_n_folds}{An integer of folds for cross-validation. 'Expressed' and
'unexpressed' training transcripts will be stratified
for calculating AUC.
Default: 10}

\item{nthreads}{An integer defining the number of threads to-be-used.
An ideal number would be equal to the maximum one of
cv_n_folds and number of input BED files. Default: 1}
}
\examples{

fbeds = c( system.file('extdata/bed/GM12878POLR2AphosphoS5Rep1.bed',
                       package='pram'),
           system.file('extdata/bed/GM12878POLR2AphosphoS5Rep2.bed',
                       package='pram') )

ftpms = c( system.file('extdata/rsem/GM12878Rep1_training.isoforms.results',
                       package='pram'),
           system.file('extdata/rsem/GM12878Rep2_training.isoforms.results',
                       package='pram') )

fgtf_training = system.file('extdata/gtf/rf_training.gtf', package='pram')
fgtf_testing  = system.file('extdata/gtf/rf_testing.gtf',  package='pram')

fgtf_out = tempfile(fileext='.gtf')

screenModel(fbeds, ftpms, fgtf_training, fgtf_testing, fgtf_out)


}
