% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenModel.R
\name{screenModel}
\alias{screenModel}
\title{Screen transcript models by a random forest classifier trained on known
transcript}
\usage{
screenModel(fbeds, ftpms, fgtf_training, fgtf_testing, fgtf_out,
  tmpdir = NULL, expr_min_tpm = 1, cv_n_folds = 10, nthreads = 1)
}
\arguments{
\item{fbeds}{A vector of BED file names for histone marks and/or RNA Pol II
ChIP-seq alignments.  Each BED file is required to have at least
four columns describing chromosome names, starting and
ending positions, and strand.}

\item{ftpms}{A vector of RSEM quantification result file names at the
isoform level.  It is required to have two columns named
'transcript_id' and 'TPM' to define the abundances of
training transcripts.  This file is the one with a suffix
'.isoforms.results' generated by RSEM.}

\item{fgtf_training}{Name of a GTF file that defines the genomic coordinates and features
for all training transcripts.  This file is required to have 'exon'
in the feature column (the 3rd column) and 'transcript_id' entry in
the attribute column (the 9th column).}

\item{fgtf_testing}{Name of a GTF file defining the genomic coordiantes and features
for transcript models to be screened.  GTF files generated by
PRAM's buildModel() function can be used here directly.}

\item{fgtf_out}{Name of a GTF file to save transcripts selected by the classifier}

\item{tmpdir}{Name of a temporary directory to save four files:
\itemize{
    \item 'rf.rda' for an R random forest classifier
    \item 'roc_pr.pdf' plots ROC and precision-recall curves
          from cross validation
    \item 'cv.tsv' saved the prediction results on training
          transcript from cross validation.
    \item 'pred.tsv' saved the prediction results testing transcripts
}
File with the same name in this folder will be overwritten.
Default: NULL}

\item{expr_min_tpm}{A numeric value of minimium TPM for a transcript to be defined as
expressed.  An 'expressed' or 'unexpressed'
transcript must have TPM >= or < this cutoff in all
supplied RSEM quantification files (ftpms),
respectively.  Otherwise, a transcript's expression
status will be defined as 'indeterminate'.  Transcripts with
'indeterminate' expression status will be omitted from the
training and output files.
Default: 1}

\item{cv_n_folds}{An integer of folds for cross validation. 'Expressed' and
'unexpressed' training transcripts will be stratified
for calculating AUC.
Default: 10}

\item{nthreads}{An integer defining the number of threads to-be-used.
Default: 1}
}
\description{
Screen transcript models by a random forest classifier trained on known
transcript
}
