% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModelFilter.R
\name{trainModelFilter}
\alias{trainModelFilter}
\title{Train a random forest classifier to filter transcript models}
\usage{
trainModelFilter(fbeds, ftpm, fgtf, expr_min_tpm = 1, nthreads = 1)
}
\arguments{
\item{fbeds}{A character vector of BED files for histone marks and/or RNA
Pol II ChIP-seq signals in RPKM.  They are required to have
at least three columns describing the names of chromosome,
starting and ending positions, and strand.}

\item{ftpm}{A character string of an RSEM quantification result at the
isoform level.  It is required to have two columns named
'transcript_id' and 'TPM' to define the abundances of
training transcripts.  This file is the one with a suffix
'.isoforms.results' generated by RSEM.}

\item{fgtf}{A string of a GTF file defining the genomic coordinates for
all training transcripts}

\item{expr_min_tpm}{Minimium TPM a transcript is considerred to be
expressed.  Default: 1.}

\item{nthreads}{an integer defining the number of threads to-be-used.
Default: 1}
}
\value{
A random forest classifier object from package randomForest.
}
\description{
Train a random forest classifier to filter transcript models
}
